InstallmentsInfoFactory
